Added in some useful middleware: IsAdmin, IsSuperuser, IsSuperuserOrGuest.
Added in constants to translate between role_id and the roles string literal.

Added login into the HomeController that checks if a user is an admin, superuser or customer and shows the correct view.

Added in a navbar for admin, superuser.

Added in a create admin page where superusers can create new admins.